Update our EXAMPLE Makedefs for modern homebrew on OSX.
Homebrew prefix paths diverge based on the underlying architecture. It appears
that the M1 prefix is `/opt/homebrew` while the x86 prefix is `/usr/local` (or
at least is sometimes `/usr/local`).

In either case, we can ask homebrew for its prefix using `homebrew config`,
which prints out a bunch of information. This updates our Makedefs.EXMAMPLE file
to use a Make `shell` command to extract the right path in an
architecture-independent manner.